+++ packages/app/obojobo-document-json-parser/src/break-node-parser/break-node-parser.js
@@ -1,0 +1,9 @@
+const processAttrs = require('../process-attrs')
+
+const breakNodeParser = node => {
+    const attrs = processAttrs(node.content, [])
+
+    return `<hr${attrs} />`
+}
+
+module.exports = breakNodeParser
