--- packages/app/obojobo-document-json-parser/src/breakNodeParser/breakNodeParser.js
@@ -1,9 +1,0 @@
-const processAttrs = require('../processAttrs')
-
-const breakNodeParser = node => {
-    const attrs = processAttrs(node.content, [])
-
-    return `<hr${attrs} />`
-}
-
-module.exports = breakNodeParser
