+++ packages/app/obojobo-document-json-parser/src/code-node-parser/code-node-parser.js
@@ -1,0 +1,14 @@
+const textGroupParser = require('../text-group-parser')
+
+const codeNodeParser = node => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const textGroupXML = textGroupParser(node.content.textGroup)
+
+    return (
+        `<Code${id}>` +
+        textGroupXML +
+        `</Code>`
+    )
+}
+
+module.exports = codeNodeParser
