--- packages/app/obojobo-document-json-parser/src/codeNodeParser/codeNodeParser.js
@@ -1,14 +1,0 @@
-const textGroupParser = require('../textGroupParser')
-
-const codeNodeParser = node => {
-    const id = node.id ? ` id="${node.id}"` : ''
-    const textGroupXML = textGroupParser(node.content.textGroup)
-
-    return (
-        `<Code${id}>` +
-        textGroupXML +
-        `</Code>`
-    )
-}
-
-module.exports = codeNodeParser
