+++ packages/app/obojobo-document-json-parser/src/heading-node-parser/heading-node-parser.js
@@ -1,0 +1,16 @@
+const textGroupParser = require('../text-group-parser')
+
+const headingNodeParser = node => {
+    const id = node.id ? ` id="${node.id}"` : ''
+
+    const headingLevel = node.content.headingLevel ? `headingLevel="${node.content.headingLevel}"` : ''
+    const textGroupXML = textGroupParser(node.content.textGroup)
+
+    return (
+        `<Heading ${headingLevel}${id}>` +
+        textGroupXML +
+        `</Heading>`
+    )
+}
+
+module.exports = headingNodeParser
