--- packages/app/obojobo-document-json-parser/src/headingNodeParser/headingNodeParser.js
@@ -1,16 +1,0 @@
-const textGroupParser = require('../textGroupParser')
-
-const headingNodeParser = node => {
-    const id = node.id ? ` id="${node.id}"` : ''
-
-    const headingLevel = node.content.headingLevel ? `headingLevel="${node.content.headingLevel}"` : ''
-    const textGroupXML = textGroupParser(node.content.textGroup)
-
-    return (
-        `<Heading ${headingLevel}${id}>` +
-        textGroupXML +
-        `</Heading>`
-    )
-}
-
-module.exports = headingNodeParser
