+++ packages/app/obojobo-document-json-parser/src/html-node-parser/html-node-parser.js
@@ -1,0 +1,10 @@
+const processAttrs = require('../process-attrs')
+
+const htmlNodeParser = node => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const attrs = processAttrs(node.content, [])
+
+    return `<HTML${attrs}${id} />`
+}
+
+module.exports = htmlNodeParser
