--- packages/app/obojobo-document-json-parser/src/htmlNodeParser/htmlNodeParser.js
@@ -1,10 +1,0 @@
-const processAttrs = require('../processAttrs')
-
-const htmlNodeParser = node => {
-    const id = node.id ? ` id="${node.id}"` : ''
-    const attrs = processAttrs(node.content, [])
-
-    return `<HTML${attrs}${id} />`
-}
-
-module.exports = htmlNodeParser
