--- packages/app/obojobo-document-json-parser/src/iFrameNodeParser/iFrameNodeParser.js
@@ -1,10 +1,0 @@
-const processAttrs = require('../processAttrs')
-
-const iFrameNodeParser = node => {
-    const id = node.id ? ` id="${node.id}"` : ''
-    const attrs = processAttrs(node.content, [])
-
-    return `<IFrame${attrs}${id} />`
-}
-
-module.exports = iFrameNodeParser
