+++ packages/app/obojobo-document-json-parser/src/iframe-node-parser/iframe-node-parser.js
@@ -1,0 +1,10 @@
+const processAttrs = require('../process-attrs')
+
+const iFrameNodeParser = node => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const attrs = processAttrs(node.content, [])
+
+    return `<IFrame${attrs}${id} />`
+}
+
+module.exports = iFrameNodeParser
