+++ packages/app/obojobo-document-json-parser/src/math-equation-node-parser/math-equation-node-parser.js
@@ -1,0 +1,10 @@
+const processAttrs = require('../process-attrs')
+
+const mathEquationNodeParser = node => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const attrs = processAttrs(node.content, [])
+
+    return `<MathEquation${attrs}${id} />`
+}
+
+module.exports = mathEquationNodeParser
