--- packages/app/obojobo-document-json-parser/src/mathEquationNodeParser/mathEquationNodeParser.js
@@ -1,10 +1,0 @@
-const processAttrs = require('../processAttrs')
-
-const mathEquationNodeParser = node => {
-    const id = node.id ? ` id="${node.id}"` : ''
-    const attrs = processAttrs(node.content, [])
-
-    return `<MathEquation${attrs}${id} />`
-}
-
-module.exports = mathEquationNodeParser
