+++ packages/app/obojobo-document-json-parser/src/mc-answer-node-parser/mc-answer-node-parser.js
