+++ packages/app/obojobo-document-json-parser/src/mc-assessment-node-parser/mcAssessmentNodeParser.js
@@ -1,0 +1,14 @@
+const processAttrs = require('../process-attrs')
+
+const mcAssessmentNodeParser = (node, childrenParser) => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const attrs = processAttrs(node.content, [])
+
+    return (
+        `<MCAssessment${attrs}${id}>` +
+        childrenParser(node.children) +
+        `</MCAssessment>`
+    )
+}
+
+module.exports = mcAssessmentNodeParser
