+++ packages/app/obojobo-document-json-parser/src/mc-choice-node-parser/mc-choice-node-parser.js
@@ -1,0 +1,14 @@
+const processAttrs = require('../process-attrs')
+
+const mcChoiceNodeParser = (node, childrenParser) => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const attrs = processAttrs(node.content, [])
+
+    return (
+        `<MCChoice${attrs}${id}>` +
+        childrenParser(node.children) +
+        `</MCChoice>`
+    )
+}
+
+module.exports = mcChoiceNodeParser
