--- packages/app/obojobo-document-json-parser/src/mcAssessmentNodeParser/mcAssessmentNodeParser.js
@@ -1,14 +1,0 @@
-const processAttrs = require('../processAttrs')
-
-const mcAssessmentNodeParser = (node, childrenParser) => {
-    const id = node.id ? ` id="${node.id}"` : ''
-    const attrs = processAttrs(node.content, [])
-
-    return (
-        `<MCAssessment${attrs}${id}>` +
-        childrenParser(node.children) +
-        `</MCAssessment>`
-    )
-}
-
-module.exports = mcAssessmentNodeParser
