+++ packages/app/obojobo-document-json-parser/src/process-attrs.js
@@ -1,0 +1,13 @@
+const xmlEncode = require('./xml-encode')
+
+const processAttrs = (content, blackListedAttributes) => {
+    let attrs = ''
+    for (const attr in content) {
+        if (content[attr] === null || blackListedAttributes.includes(attr)) continue;
+        attrs += ` ${attr}="${xmlEncode(content[attr])}"`
+    }
+
+    return attrs
+}
+
+module.exports = processAttrs
