--- packages/app/obojobo-document-json-parser/src/processAttrs.js
@@ -1,13 +1,0 @@
-const xmlEncode = require('./xmlEncode')
-
-const processAttrs = (content, blackListedAttributes) => {
-    let attrs = ''
-    for (const attr in content) {
-        if (content[attr] === null || blackListedAttributes.includes(attr)) continue;
-        attrs += ` ${attr}="${xmlEncode(content[attr])}"`
-    }
-
-    return attrs
-}
-
-module.exports = processAttrs
