+++ packages/app/obojobo-document-json-parser/src/question-bank-node-parser/question-bank-node-parser.js
@@ -1,0 +1,14 @@
+const processAttrs = require('../process-attrs')
+
+const questionBankNodeParser = (node, childrenParser) => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const attrs = processAttrs(node.content, [])
+
+    return (
+        `<QuestionBank${attrs}${id}>` +
+        childrenParser(node.children) +
+        `</QuestionBank>`
+    )
+}
+
+module.exports = questionBankNodeParser
