--- packages/app/obojobo-document-json-parser/src/questionBankNodeParser/questionBankNodeParser.js
@@ -1,14 +1,0 @@
-const processAttrs = require('../processAttrs')
-
-const questionBankNodeParser = (node, childrenParser) => {
-    const id = node.id ? ` id="${node.id}"` : ''
-    const attrs = processAttrs(node.content, [])
-
-    return (
-        `<QuestionBank${attrs}${id}>` +
-        childrenParser(node.children) +
-        `</QuestionBank>`
-    )
-}
-
-module.exports = questionBankNodeParser
