+++ packages/app/obojobo-document-json-parser/src/text-node-parser/text-node-parser.js
@@ -1,0 +1,14 @@
+const textGroupParser = require('../text-group-parser')
+
+const textNodeParser = node => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const textGroupXML = textGroupParser(node.content.textGroup)
+
+    return (
+        `<Text${id}>` +
+        textGroupXML +
+        `</Text>`
+    )
+}
+
+module.exports = textNodeParser
