--- packages/app/obojobo-document-json-parser/src/textNodeParser/textNodeParser.js
@@ -1,14 +1,0 @@
-const textGroupParser = require('../textGroupParser')
-
-const textNodeParser = node => {
-    const id = node.id ? ` id="${node.id}"` : ''
-    const textGroupXML = textGroupParser(node.content.textGroup)
-
-    return (
-        `<Text${id}>` +
-        textGroupXML +
-        `</Text>`
-    )
-}
-
-module.exports = textNodeParser
