+++ packages/app/obojobo-document-json-parser/src/xml-encode.js
