+++ packages/app/obojobo-document-json-parser/src/youtube-node-parser/youtube-node-parser.js
@@ -1,0 +1,10 @@
+const xmlEncode = require('../xml-encode')
+
+const youTubeNodeParser = node => {
+    const id = node.id ? ` id="${node.id}"` : ''
+    const videoId = xmlEncode(node.content.videoId)
+
+    return `<YouTube videoId="${videoId}"${id}/>`
+}
+
+module.exports = youTubeNodeParser
